services/gmail-mcp/requirements.txt
mcp>=1.9,<2.0
httpx>=0.28,<1.0
pyjwt>=2.10,<3.0
google-api-python-client>=2.160,<3.0
google-auth>=2.38,<3.0
google-auth-oauthlib>=1.2,<2.0
uvicorn[standard]>=0.35,<1.0
